+++ edit_schema_generator/lib/edit_schema_generator.dart
@@ -1,4 +1,5 @@
 library edit_schema_generator;
 
 export 'src/edit_schema.dart';
+export 'src/lens.dart';
 export 'src/source_gen/edit_schema_annotations.dart';