+++ edit_schema_generator/test/schema/fleet_single_tree_fixture.dart
@@ -1,7 +1,6 @@
 import 'package:edit_schema_generator/edit_schema_generator.dart';
 
 import 'fleet_test_models.dart';
-import 'test_lens.dart';
 
 part 'fleet_single_tree_fixture.g.dart';